hyperv: don't error out if cleanup fails (#670)

+++ Libraries/HyperV.psm1
@@ -175,35 +175,38 @@ Function Delete-HyperVGroup([string]$HyperVGroupName, [string]$HyperVHost, $Setu
 
     $cleanupDone = 0
     $vmGroup.VMMembers | ForEach-Object {
-        $vm = $_
+        $vm = Get-VM -Name $_.Name -ComputerName $HyperVHost
+
         Write-LogInfo "Stop-VM -Name $($vm.Name) -Force -TurnOff"
         Stop-VM -Name $vm.Name -Force -TurnOff -ComputerName $HyperVHost
-        $snapshots = Get-VMSnapshot -VMName $vm.Name -ComputerName $HyperVHost
-        if ($snapshots.Count -gt 1 -and ($snapshots.Name -join "") -imatch "fail") {
-            Write-LogWarn "VM $($vm.Name) cannot be cleaned up as it has failed test cases snapshots."
-            $cleanupDone--
-            return
+        try {
+            Wait-VMState -VMName $vm.Name -VMState "Off" -RetryInterval 3 `
+                -HvServer $HyperVHost
+            Wait-VMStatus -VMName $vm.Name -VMStatus "Operating Normally" -RetryInterval 3 `
+                -HvServer $HyperVHost
+        } catch {
+            return $false
         }
-        $retriesCleanSnapshots = 0
-        $maxRetriesCleanSnapshots = 3
-        while ($retriesCleanSnapshots -lt $maxRetriesCleanSnapshots) {
-            Remove-VMSnapshot -VMName $vm.Name -ComputerName $HyperVHost `
-                -IncludeAllChildCheckpoints -Confirm:$false -ErrorAction SilentlyContinue
-            if ($?) {
-                Write-LogWarn ("Failed to remove snapshots for VM {0}. Retrying..." -f @($vm.Name))
-                $retriesCleanSnapshots++
-            } else {
-                break
+
+        # Note(v-advlad): Need to remove also the parents of the .avhdx (snapshots)
+        $hardDiskPath = @()
+        $vm.HardDrives | ForEach-Object {
+            $hardDiskPath += $_.Path
+            if ($_.Path -match ".avhdx") {
+                $snapshotParent = Get-VHD $_.Path -ComputerName $HyperVHost
+                if ($snapshotParent -and $snapshotParent.ParentPath) {
+                    $hardDiskPath += $snapshotParent.ParentPath
+                }
             }
         }
-        Wait-VMStatus -VMName $vm.Name -VMStatus "Operating Normally" -RetryInterval 2 `
-            -HvServer $HyperVHost
-        $vm = Get-VM -Name $vm.Name -ComputerName $HyperVHost
-        $vm.HardDrives | ForEach-Object {
-            $vhdPath = $_.Path
+
+        $hardDiskPath | ForEach-Object {
+            $vhdPath = $_
             $invokeCommandParams = @{
                 "ScriptBlock" = {
-                    Remove-Item -Path $args[0] -Force
+                    if ((Test-Path $args[0])) {
+                        Remove-Item -Path $args[0] -Force
+                    }
                 };
                 "ArgumentList" = $vhdPath;
             }
